+++ components/src/main/java/uk/gov/hmrc/components/organism/summary/SummaryRowView.kt
@@ -21,7 +21,10 @@ import android.util.AttributeSet
 import android.view.LayoutInflater
 import android.view.View
 import android.widget.RelativeLayout
+import androidx.annotation.ColorRes
+import androidx.annotation.DrawableRes
 import androidx.annotation.StyleRes
+import androidx.core.content.ContextCompat
 import androidx.core.widget.TextViewCompat
 import uk.gov.hmrc.components.R
 import uk.gov.hmrc.components.databinding.ComponentSummaryRowBinding
@@ -65,6 +68,9 @@ class SummaryRowView @JvmOverloads constructor(
                 R.styleable.SummaryRowView_titleTextAppearance,
                 R.style.Text_Bold
             )
+            val icon = typedArray.getResourceId(R.styleable.SummaryRowView_chevronIcon, R.drawable.components_ic_chevron_right)
+            val iconTint = typedArray.getResourceId(R.styleable.SummaryRowView_chevronIconTintColor, R.color.hmrc_black)
+
             val titleMaxLines = typedArray.getInt(R.styleable.SummaryRowView_titleMaxLines, -1)
             val accessibilityMessage = typedArray.getString(R.styleable.SummaryRowView_accessibilityMessage)
             val readerTrait = typedArray.getInt(R.styleable.SummaryRowView_readerTrait, READER_TRAIT_INFO)
@@ -73,6 +79,8 @@ class SummaryRowView @JvmOverloads constructor(
             setTitle(titleText)
             setTitleTextAppearance(titleTextAppearance)
             setTitleMaxLines(titleMaxLines)
+            setChevronIcon(icon)
+            setChevronTintColor(iconTint)
 
             accessibilityMessage?.let { message ->
                 setAccessibilityMessage(message)
@@ -125,6 +133,21 @@ class SummaryRowView @JvmOverloads constructor(
         }
     }
 
+    fun setChevronIcon(@DrawableRes icon: Int) {
+        if (icon != NO_CHEVRON_ICON) {
+            binding.imageChevron.setImageResource(icon)
+        }
+    }
+
+    fun setChevronTintColor(@ColorRes iconTint: Int) {
+        if (iconTint != NO_CHEVRON_ICON_TINT) {
+            binding.imageChevron.setColorFilter(
+                ContextCompat.getColor(context, iconTint),
+                android.graphics.PorterDuff.Mode.SRC_IN
+            )
+        }
+    }
+
     fun setOnClickListener(clickHandler: () -> Unit) {
         isClickable = true
         isFocusable = true
@@ -159,5 +182,7 @@ class SummaryRowView @JvmOverloads constructor(
     companion object {
         const val READER_TRAIT_INFO = 1
         const val READER_TRAIT_SIMPLE = 2
+        const val NO_CHEVRON_ICON_TINT: Int = -1
+        const val NO_CHEVRON_ICON: Int = -1
     }
 }