+++ sample/src/main/java/uk/gov/hmrc/components/sample/organisms/SummaryRowFragment.kt
@@ -134,6 +134,14 @@ class SummaryRowFragment : BaseComponentsFragment() {
             setRows(arrayListOf(example5Row1, example5Row2))
             setOnClickListener { onCtaPressed() }
         }
+
+        val example6Row1 = MultiColumnRowView(requireContext())
+        example6Row1.setText(text1 = getString(R.string.long_text))
+        binding.summaryRowExample6.apply {
+            setRows(arrayListOf(example6Row1))
+            setOnClickListener { onCtaPressed() }
+            setChevronTintColor(R.color.hmrc_blue)
+        }
     }
 
     private fun onCtaPressed() {